* master:
  Update POT files using the production database
  Update translations from Transifex
  MBS-12079: JSON-LD for releases with a lot of [silence] is too big (#2331)
  MBS-12034: Require confirmation to enter a "Disc n" medium title (#2312)
  MBS-11917: New report: Releases with "Disc n" medium titles (#2338)
  Try to eradicate 5xx errors during deployment (#2297)
  MBS-11991: Add links for uses of artist credits in entity editor (#2304)
  MBS-12050: Add "edit note does not include" to edit search (#2318)
  MBS-12077: Also block useless edit notes on the release editor (#2334)
  Add test for viewing subscriptions
  MBS-12055: Allow admins to see private editor-editor subscriptions
  MBS-12001: Use allowNew for recording links when adding a new track (#2311)
  MBS-12084: Use URL to build the details API links (#2333)
  MBS-12057: Show entity type on sidebar "No collections!" message (#2319)
  MBS-12060: Block google.com/search URLs on URL relationships (#2323)
  Add a Selenium test for CD TOC lookup/attach
  MBS-12087: Fix TT TaggerIcon call
  MBS-12059: Autoselect YesAsia as "purchase for mail-order" + sidebar
  MBS-12063: Autoselect Target as "purchase for mail-order" + sidebar
  MBS-12064: Add jazzmusicarchives.com to the database whitelist
  MBS-12021: Properly override primary aliases in code
  Avoid concatenation warning for unset locale

+++ lib/MusicBrainz/Server/Controller/User/SubscriptionsRole.pm
@@ -28,7 +28,10 @@ role {
 
         my $user = $c->stash->{user};
 
-        if (!defined $c->user || $c->user->id != $user->id) {
+        # Admins should have access to editor-editor subscriptions (MBS-12055)
+        if (!defined $c->user || $c->user->id != $user->id && !(
+            $type eq 'editor' && $c->user->is_account_admin
+        )) {
             $c->model('Editor')->load_preferences($user);
             $c->detach('/error_403')
                 unless $user->preferences->public_subscriptions;
